+++ packages/orq-rc/src/orq_ai_sdk/guardrailrules.py
@@ -315,7 +315,7 @@ class GuardrailRules(BaseSDK):
             return unmarshal_json_response(
                 models.GuardrailRuleCreateResponseBody, http_res
             )
-        if utils.match_response(http_res, ["400", "4XX"], "*"):
+        if utils.match_response(http_res, ["400", "409", "4XX"], "*"):
             http_res_text = utils.stream_to_text(http_res)
             raise models.APIError("API error occurred", http_res, http_res_text)
         if utils.match_response(http_res, "5XX", "*"):
@@ -433,7 +433,7 @@ class GuardrailRules(BaseSDK):
             return unmarshal_json_response(
                 models.GuardrailRuleCreateResponseBody, http_res
             )
-        if utils.match_response(http_res, ["400", "4XX"], "*"):
+        if utils.match_response(http_res, ["400", "409", "4XX"], "*"):
             http_res_text = await utils.stream_to_text_async(http_res)
             raise models.APIError("API error occurred", http_res, http_res_text)
         if utils.match_response(http_res, "5XX", "*"):
@@ -909,7 +909,7 @@ class GuardrailRules(BaseSDK):
             return unmarshal_json_response(
                 models.GuardrailRuleUpdateResponseBody, http_res
             )
-        if utils.match_response(http_res, ["400", "404", "4XX"], "*"):
+        if utils.match_response(http_res, ["400", "404", "409", "4XX"], "*"):
             http_res_text = utils.stream_to_text(http_res)
             raise models.APIError("API error occurred", http_res, http_res_text)
         if utils.match_response(http_res, "5XX", "*"):
@@ -1033,7 +1033,7 @@ class GuardrailRules(BaseSDK):
             return unmarshal_json_response(
                 models.GuardrailRuleUpdateResponseBody, http_res
             )
-        if utils.match_response(http_res, ["400", "404", "4XX"], "*"):
+        if utils.match_response(http_res, ["400", "404", "409", "4XX"], "*"):
             http_res_text = await utils.stream_to_text_async(http_res)
             raise models.APIError("API error occurred", http_res, http_res_text)
         if utils.match_response(http_res, "5XX", "*"):