+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KPSXXF499ZRKWG45AX72VXCK"
+        "01KPT6B8CS3C7MAVG9WKWT27BS"
     )
     r"""The unique identifier of the data source"""