+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KPSXXF4BQCQ12SPWVXX4NWNR"
+        "01KPT6B8CVADQSH7FQB9HK2D1Z"
     )
     r"""The unique identifier of the data source"""